requirements.txt
# requirements.txt
pandas
requests
Pillow
numpy
tqdm
dvc
scikit-learn
mlflow
PyYAML